[ANA-1840] Scientific Notation Error Handling (#70)
* Add logic cast part 1

* Add logic cast part 2

* Add logic cast part 3

* comment code test

* Add unit test code in GV Details

* Add unit test code in LP Summary and Details

* Bump up version 0.4.18 to 0.7.0

+++ aggr-ana-bs-summary/src/main/scala/com/eci/anaplan/bs/summary/aggregations/joiners/BusSummary.scala
@@ -2,6 +2,7 @@ package com.eci.anaplan.bs.summary.aggregations.joiners
 
 import com.eci.anaplan.bs.summary.aggregations.constructors.BusSummaryDf
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types.{DecimalType, IntegerType}
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
 import javax.inject.{Inject, Singleton}
@@ -21,8 +22,8 @@ class BusSummary @Inject()(spark: SparkSession,
         $"product",
         coalesce($"payment_channel",lit("None")).as("payment_channel"))
       .agg(
-        coalesce(countDistinct($"booking_id"),lit(0)).as("no_of_transactions"),
-        coalesce(sum($"coupon_code_result"),lit(0)).as("no_of_coupon")
+        coalesce(countDistinct($"booking_id"),lit(0)).cast(IntegerType).as("no_of_transactions"),
+        coalesce(sum($"coupon_code_result"),lit(0)).cast(IntegerType).as("no_of_coupon")
       )
       .select(
         $"*"
